create classes for impendace
